refactor billing display for isolated tests
