+++ client/dashboard/src/pages/billing/Billing.test.tsx
@@ -2,7 +2,10 @@ import { cleanup, render, screen } from "@testing-library/react";
 import type { TierLimits } from "@gram/client/models/components/tierlimits.js";
 import type { UsageTiers } from "@gram/client/models/components/usagetiers.js";
 import { afterEach, describe, expect, it } from "vitest";
-import { getChatCreditEntitlement, TierIncludedItems } from "./Billing";
+import {
+  getChatCreditEntitlement,
+  TierIncludedItems,
+} from "./billing-entitlements";
 
 afterEach(cleanup);
